move test, add more tests
