+++ tests/Console/ConfigurationResolverTest.php
@@ -16,20 +16,15 @@ use PhpCsFixer\Config;
 use PhpCsFixer\Console\Command\FixCommand;
 use PhpCsFixer\Console\ConfigurationResolver;
 use PhpCsFixer\Finder;
-use PhpCsFixer\RuleSet;
-use PhpCsFixer\Test\AccessibleObject;
 use Symfony\Component\Console\Output\OutputInterface;
 
 /**
  * @author Katsuhiro Ogawa <ko.fivestar@gmail.com>
  * @author Dariusz Rumiński <dariusz.ruminski@gmail.com>
- * @author SpacePossum
  *
  * @internal
  *
  * @covers \PhpCsFixer\Console\ConfigurationResolver
- * @covers \PhpCsFixer\RuleSet::resolveSet
- * @covers \PhpCsFixer\RuleSet::resolveSubset
  */
 final class ConfigurationResolverTest extends \PHPUnit_Framework_TestCase
 {
@@ -948,148 +943,6 @@ final class ConfigurationResolverTest extends \PHPUnit_Framework_TestCase
         );
     }
 
-    /**
-     * @param array $expected
-     * @param array $rules
-     *
-     * @dataProvider provideResolveRulesCases
-     */
-    public function testResolveRules(array $expected, array $rules)
-    {
-        $ruleSet = $this->createRuleSetToTestWith($rules);
-
-        $this->assertSameRules($expected, $ruleSet->getRules());
-    }
-
-    public function provideResolveRulesCases()
-    {
-        return array(
-            '@Foo + C\' -D' => array(
-                array('A' => true, 'B' => true, 'C' => 56),
-                array('@Foo' => true, 'C' => 56, 'D' => false),
-            ),
-            '@Foo + @Bar' => array(
-                array('A' => true, 'B' => true, 'D' => 34, 'E' => true),
-                array('@Foo' => true, '@Bar' => true),
-            ),
-            '@Foo - @Bar' => array(
-                array('B' => true),
-                array('@Foo' => true, '@Bar' => false),
-            ),
-            '@A - @E (set in set)' => array(
-                array('AA' => true), // 'AB' => false, 'AC' => false
-                array('@A' => true, '@E' => false),
-            ),
-            '@A + @E (set in set)' => array(
-                array('AA' => true, 'AB' => '_AB', 'AC' => 'b', 'Z' => true),
-                array('@A' => true, '@E' => true),
-            ),
-            '@E + @A (set in set) + rule override' => array(
-                array('AC' => 'd', 'AB' => true, 'Z' => true, 'AA' => true),
-                array('@E' => true, '@A' => true, 'AC' => 'd'),
-            ),
-            'nest single set' => array(
-                array('AC' => 'b', 'AB' => '_AB', 'Z' => 'E'),
-                array('@F' => true),
-            ),
-            'Set reconfigure rule in other set, reconfigure rule.' => array(
-                array(
-                    'AA' => true,
-                    'AB' => true,
-                    'AC' => 'abc',
-                ),
-                array(
-                    '@A' => true,
-                    '@D' => true,
-                    'AC' => 'abc',
-                ),
-            ),
-            'Set reconfigure rule in other set.' => array(
-                array(
-                    'AA' => true,
-                    'AB' => true,
-                    'AC' => 'b',
-                ),
-                array(
-                    '@A' => true,
-                    '@D' => true,
-                ),
-            ),
-            'Set minus two sets minus rule' => array(
-                array(
-                    'AB' => true,
-                ),
-                array(
-                    '@A' => true,
-                    '@B' => false,
-                    '@C' => false,
-                    'AC' => false,
-                ),
-            ),
-            'Set minus two sets' => array(
-                array(
-                    'AB' => true,
-                    'AC' => 'a',
-                ),
-                array(
-                    '@A' => true,
-                    '@B' => false,
-                    '@C' => false,
-                ),
-            ),
-            'Set minus rule test.' => array(
-                array(
-                    'AA' => true,
-                    'AC' => 'a',
-                ),
-                array(
-                    '@A' => true,
-                    'AB' => false,
-                ),
-            ),
-            'Set minus set test.' => array(
-                array(
-                    'AB' => true,
-                    'AC' => 'a',
-                ),
-                array(
-                    '@A' => true,
-                    '@B' => false,
-                ),
-            ),
-            'Set to rules test.' => array(
-                array(
-                    'AA' => true,
-                    'AB' => true,
-                    'AC' => 'a',
-                ),
-                array(
-                    '@A' => true,
-                ),
-            ),
-            '@A - @C' => array(
-                array(
-                    'AB' => true,
-                    'AC' => 'a',
-                ),
-                array(
-                    '@A' => true,
-                    '@C' => false,
-                ),
-            ),
-            '@A - @D' => array(
-                array(
-                    'AA' => true,
-                    'AB' => true,
-                ),
-                array(
-                    '@A' => true,
-                    '@D' => false,
-                ),
-            ),
-        );
-    }
-
     public function testResolveRulesWithUnknownRules()
     {
         $this->setExpectedException(
@@ -1183,45 +1036,6 @@ final class ConfigurationResolverTest extends \PHPUnit_Framework_TestCase
         $this->assertFalse($resolver->getUsingCache());
         $this->assertNull($resolver->getCacheFile());
         $this->assertSame('xml', $resolver->getReporter()->getFormat());
-    }
-
-    private function createRuleSetToTestWith(array $rules)
-    {
-        static $testSet = array(
-            '@A' => array(
-                'AA' => true,
-                'AB' => true,
-                'AC' => 'a',
-            ),
-            '@B' => array(
-                'AA' => true,
-            ),
-            '@C' => array(
-                'AA' => false,
-            ),
-            '@D' => array(
-                'AC' => 'b',
-            ),
-            '@E' => array(
-                '@D' => true,
-                'AB' => '_AB',
-                'Z' => true,
-            ),
-            '@F' => array(
-                '@E' => true,
-                'Z' => 'E',
-            ),
-            '@Foo' => array('A' => true, 'B' => true, 'C' => true, 'D' => 12),
-            '@Bar' => array('A' => true, 'C' => false, 'D' => 34, 'E' => true, 'F' => false),
-        );
-
-        $ruleSet = new RuleSet();
-        $reflection = new AccessibleObject($ruleSet);
-        $reflection->setDefinitions = $testSet;
-        $reflection->set = $rules;
-        $reflection->resolveSet();
-
-        return $ruleSet;
     }
 
     private function assertSameRules(array $expected, array $actual, $message = '')
